fix(studio): extract visual fidelity helper
